# WARNING: head commit changed in the meantime
Merge branch 'main' of https://github.com/DDemmer1/26-10-2022 into main

+++ src/de/demmer/dennis/Application.java
@@ -34,8 +34,9 @@ public class Application {
 	}
 	
 	
-	static double durchZwei(int zahl) {
-		return zahl /2D;
+	static double durchZwei(double zahl) {
+		
+		return zahl / 2;
 	}
